Create Palindrome Number
Intuition
The intuition behind this problem is to recognize that a palindrome number reads the same forward and backward. For a given integer x, if it is negative, it cannot be a palindrome since the negative sign will only appear on the left, making it impossible to mirror. For positive integers, we can reverse the number and check if it matches the original number to determine if it's a palindrome.

Approach
If the number x is negative, return false immediately as negative numbers cannot be palindromes.
For positive numbers:
Initialize variables x_reverse to store the reversed number and n as a copy of x.
Use a loop to reverse the digits of n. In each iteration, extract the last digit (rem), multiply the current x_reverse by 10 and add rem to it. Then, divide n by 10 to discard the last digit.
Once the loop completes, x_reverse will contain the reversed number.
Finally, compare the reversed number x_reverse with the original number x. If they are equal, return true, indicating that the number is a palindrome. Otherwise, return false.
Complexity
Time complexity: O(log₁₀(n)), where n is the value of the input number x. This is because the number of digits in the number determines how many times we loop to reverse the digits. The number of digits is proportional to the logarithm of n to the base 10.

Space complexity: O(1), as we only use a constant amount of extra space (a few integer variables) to store the reversed number and the remainder during the process.